Allow react_component to accept JSON strings
Rails ships with a useful JSON builder that makes rendering to strings
of JSON via templates very easy, but rendering from objects to json
not quite so easy. It's used by a lot of Rails apps as the default way
to serialize objects to JSON.

This allows render_component to take a string as the props argument,
and instead of calling #to_json on it, just pass it straight through
to the renderer.

Basically, this allows syntax like the following:

```ruby
react_component(
  'Todo',
  render(template: 'todos/show.json.jbuilder', locals: {todo: @todo}),
  {prerender: true}
)
```