Add support for saving references to components
When rendering a component using the react_component view helper,
pass `reference: 'myReferenceName'` in the options hash to make
react_ujs store a reference to the rendered component at
`window.myReferenceName`. This is useful for scripts that need to
make use of a component after a page has loaded, without needing
to re-render an extra time.