Lower the coffee-script-source version specifier
There is an issue with the latest version of `coffee-script-source`
outlined in https://github.com/jashkenas/coffeescript/issues/382.

The `react-rails` gemspec specifies a `~>1.9` version of
`coffee-script-source`, which makes it impossible to back out of
its most recent version in order to avoid the errors (for example,
if you want to specify `~>1.8` in your app).

This change lowers the dependency version in the `react-rails`
gemspec so that a user can specify a lower working version.