BabelTransformer options on a per-file basis.
Motivation:

While trying to bring some organization to client-side code we were
exploring the usage of defining AMD modules, and requiring them with a
tool such as [almond](https://github.com/jrburke/almond). The Babel
transpiler has the ability to generate AMD modules baked in, but the
we aren't able to pass the needed `moduleId` option on a per file basis.

Solution:

Provide a way to make per-file configurations by using a lambda, proc,
or anything that responds to `call` and returns a `Hash` as the
`jsx_transform_options`.

Sidenote:

This could be particularly useful for those who want to author their
code with ES6 modules syntax. I've noticed in the issues that several
users are reaching for `browserify-rails` to solve this problem. But the
dependency on NPM, and divide between npm and sprockets leads to a lot
of questions and configration. If the end-goal is to have ES6 syntax,
almond and some AMD configuration will solve that problem without having
to add another dependency.