:arrow_down: Lock webpack-dev-server for test dummy
The `webpack-dev-server` package for the dummy test app was locked to a version that required webpack 4.  Since the dummy is created for webpack 3, this was causing the builds to fail.

In the future, we will need to add more support for the different webpack versions to make sure that the functionality is supported.