support React 18's client side render methods
With React 18, the rendering interface will be updated in order to benefit from new features. https://github.com/reactwg/react-18/discussions/5

This allows React 18 rendering to be supported while creating backwards compatibility.

For the time being, the return of the components does not change the output from the previous versions so as to not cause any breaking changes.